refactor: 접근제어자 관리를 통한 캡슐화를 위한 패키지 수정
