+++ src/main/java/webserver/RequestHandler.java
@@ -14,6 +14,7 @@ import org.slf4j.LoggerFactory;
 
 import webserver.http.request.HttpRequest;
 import webserver.http.response.HttpResponse;
+import webserver.http.response.ResponseHandler;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
