+++ src/main/java/webserver/http/response/ResponseHandler.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http.response;
 
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
@@ -8,9 +8,9 @@ import java.net.URISyntaxException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import webserver.RequestMapper;
 import webserver.http.request.HttpRequest;
 import webserver.http.request.MappedRequest;
-import webserver.http.response.HttpResponse;
 
 public class ResponseHandler {
     private static final Logger logger = LoggerFactory.getLogger(ResponseHandler.class);
